StockDashboard/Backend/Azure Functions/Functions/requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
cryptography==2.7
decorator==4.4.0
entrypoints==0.3
idna==2.8
iexfinance==0.4.2
imagesize==1.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.13.3
jeepney==0.4
jsonschema==3.0.1
keyring==18.0.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
numpy==1.17.0
numpydoc==0.9.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psutil==5.6.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypi==2.1
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.2
QtPy==1.8.0
requests==2.22.0
rope==0.14.0
SecretStorage==3.1.1
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.1
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.4
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
wurlitzer==1.0.2
